Implement QuestionCard
