+++ frontend/src/sidepanel/SidePanel.tsx
@@ -1,15 +1,14 @@
 import React, { useEffect, useState } from "react";
-import { API_HOST } from "../constants";
 import { CardProps, renderCard } from "./components";
 import { SortableList } from "../shared/components/SortableList/SortableList";
-import { useFetch, usePageInfo } from "../shared/hooks";
+import { usePageInfo } from "../shared/hooks";
 
 const SidePanelHeader: React.FC = () => <h1>AI家庭教師くん（サイドパネル）</h1>;
 
 export const SidePanel: React.FC = () => {
-  const [question, setQuestion] = useState<string>("");
+  const [, setQuestion] = useState<string>("");
   const [, setSummary] = useState<string | null>(null);
-  const [response, setResponse] = useState<string | null>(null);
+  const [, setResponse] = useState<string | null>(null);
 
   const resetStates = () => {
     setSummary(null);
@@ -18,24 +17,6 @@ export const SidePanel: React.FC = () => {
   };
 
   const pageInfo = usePageInfo(resetStates);
-
-  const {
-    data: responseData,
-    loading: isQuestionLoading,
-    fetchData: fetchAnswer,
-  } = useFetch<typeof response>(`${API_HOST}/predict`, null);
-
-  useEffect(() => {
-    if (responseData !== null) setResponse(responseData);
-  }, [responseData]);
-
-  const handleQuestion = () => {
-    if (pageInfo && question) {
-      fetchAnswer({
-        text: `「${question}」と言う質問に対して、以下の情報を踏まえて回答せよ。: ${pageInfo.content}`,
-      });
-    }
-  };
 
   const [cards, setCards] = React.useState<CardProps[]>([
     {
@@ -54,20 +35,6 @@ export const SidePanel: React.FC = () => {
   return (
     <>
       <SidePanelHeader />
-      <div>
-        <h2>質問する</h2>
-        <input
-          type="text"
-          value={question}
-          onChange={(e) => setQuestion(e.target.value)}
-          placeholder="質問を入力してください"
-        />
-        <button onClick={handleQuestion} disabled={isQuestionLoading}>
-          質問する
-        </button>
-        {isQuestionLoading && <p>考え中...</p>}
-        {response && <p>{response}</p>}
-      </div>
       <div style={{ maxWidth: 400, margin: "30px auto" }}>
         <SortableList
           items={cards}
