+++ frontend/src/sidepanel/components/Question.tsx
@@ -1,18 +1,34 @@
-import React, { useState } from "react";
+import React, { useEffect, useState } from "react";
 import { Card, CardBase } from "./base";
 import { PageInfo } from "../../types/Page";
+import { useFetch } from "../../shared/hooks";
+import { API_HOST } from "../../constants";
 
 type Props = {
   pageInfo: PageInfo | null;
   model: string;
 };
 
-export const QuestionCard: React.FC<Props> = ({}) => {
+export const QuestionCard: React.FC<Props> = ({ pageInfo }) => {
   const [question, setQuestion] = useState<string>("");
-  const [answer, setAnswer] = useState("ここに回答が入ります");
+  const [answer, setAnswer] = useState<string | null>(null);
 
-  const onClick = () => {
-    setAnswer("回答しました。");
+  const {
+    data: responseData,
+    loading: isQuestionLoading,
+    fetchData: fetchAnswer,
+  } = useFetch<typeof answer>(`${API_HOST}/predict`, null);
+
+  useEffect(() => {
+    if (responseData !== null) setAnswer(responseData);
+  }, [responseData]);
+
+  const handleQuestion = () => {
+    if (pageInfo && question) {
+      fetchAnswer({
+        text: `「${question}」と言う質問に対して、以下の情報を踏まえて回答せよ。: ${pageInfo.content}`,
+      });
+    }
   };
 
   return (
@@ -23,7 +39,7 @@ export const QuestionCard: React.FC<Props> = ({}) => {
           onChange={(e) => setQuestion(e.target.value)}
           placeholder="質問を入力してください"
         />
-      <button onClick={onClick}>質問する</button>
+      <button onClick={handleQuestion}>質問する</button>
       <div
         style={{
           height: "200px",
@@ -37,7 +53,7 @@ export const QuestionCard: React.FC<Props> = ({}) => {
           borderRadius: "4px",
         }}
       >
-        {answer}
+        {isQuestionLoading ? "考え中..." : answer ?? "ここに回答が入ります"}
       </div>
     </Card>
   );
